Remove ActorGroup, Add Tracker for cleanly locking/tracking positions
Tracker should probably be in engine.objects, because it is more closely
tied to GameObject.  Trackers can be used for non-graphical
applications, such as AI following (when the unit loses sight of
whatever it's tracking, lock the tracker and the unit moves to last-seen
position.  Unlock the tracker when the followed unit is discovered again
to continue following).