Fix Pool implementation bug
Previously, the first element created (the tail) was being set as the
next element of the last element created (the head).  This explains why
only two elements (first click) and the current position of the mouse
drag were displayed in the test screen - all of the middle elements were
removed.  Flipping the order (set the next of first created to the last
created) fixed the issue, and the PoolBuffer is responsible for
maintaing order again.