Overhaul Momentum class
This one is tracking very nicely, and the interface it exposes is very
clean.  I still want to expose the falloff parameters from the Momentum
class, but I'll add that when it becomes necessary.  In the meantime,
experimentation shows that multiplying some root of the two momentums
together gives a natural-ish net momentum.